Fix cyclic dependencies in fable-core
